Make testbed usable from other repos
1. Extract perf tests to "Scenario" funcs which are parametrized and use the
   scenarios in Test funcs in this repo. This allows the same scenarios to
   be used in other repos (such as "contrib"). The intent is to perform the same
   scenarios with other protocols defined in "contrib".

2. Introduce helper funcs like DoTestMain that make it easy to use the testbed
   from other repos.

3. Make testutils package public to allow it to be used from testbeds in other repos.

4. Fixed broken targets in makefile.
